Rename the browser tool
