+++ tools/selenium-browser/package.json
@@ -11,17 +11,23 @@
   ],
   "license": "MIT",
   "devDependencies": {
+    "@types/async-exit-hook": "^2.0.0",
+    "@types/jsdom": "^21.1.1",
     "@types/node": "^18.16.3",
     "@types/selenium-webdriver": "^4.1.14",
+    "@types/turndown": "^5.0.1",
     "npm-run-all": "^4.1.5",
-    "typescript": "^5.0.4",
     "release-it": "^15.10.0",
-    "rimraf": "^5.0.0"
+    "rimraf": "^5.0.0",
+    "typescript": "^5.0.4"
   },
   "dependencies": {
+    "async-exit-hook": "^2.0.1",
+    "cheerio": "^1.0.0-rc.12",
+    "jsdom": "22.0.0",
     "selenium-webdriver": "4.9.0",
-    "webdriver-manager": "^13.0.2",
-    "cheerio": "^1.0.0-rc.12"
+    "turndown": "^7.1.2",
+    "webdriver-manager": "^13.0.2"
   },
   "peerDependencies": {
     "agent-mimir": "*"
