+++ tools/selenium-browser/src/html-processor.ts
