--- tools/selenium-summarizer-browser/tsconfig.json
@@ -1,28 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "ES2021",
-    "allowJs": true,
-    "allowSyntheticDefaultImports": true,
-    "baseUrl": "src",
-    "declaration": true,
-    "sourceMap": true,
-    "inlineSourceMap": false,
-    "moduleResolution": "node",
-    "rootDir": "src",
-    "module": "commonjs", /* Specify what module code is generated. */
-    "outDir": "dist",
-    "lib": [
-      "ES2021",
-      "ES2022.Object",
-      "DOM"
-    ],
-    "types": [
-      "node",
-    ],
-    "esModuleInterop": true, /* Emit additional JavaScript to ease support for importing CommonJS modules. This enables 'allowSyntheticDefaultImports' for type compatibility. */
-    "forceConsistentCasingInFileNames": true, /* Ensure that casing is correct in imports. */
-    /* Type Checking */
-    "strict": true, /* Enable all strict type-checking options. */
-    "skipLibCheck": true /* Skip type checking all .d.ts files. */
-  }
-}
